Removed nock in favour of sinon mocking
